refactor: replace unused pagination.py with PaginationMixin
Delete unused PaginationHelper (0% coverage). Extract _gather_results,
_process_documents, _fetch_all_pages, _paginate_until_end, _with_save
from BaseScraper into PaginationMixin. BaseScraper now inherits from
BrowserMixin and PaginationMixin.
