+++ src/scraper/base/pagination.py
@@ -1,46 +1,130 @@
-"""Pagination helpers for BaseScraper.
+"""Pagination mixin for BaseScraper.
 
-Extracted from BaseScraper; access via ``self.paginator`` on any BaseScraper subclass.
+Provides ``_gather_results``, ``_process_documents``, ``_fetch_all_pages``,
+``_paginate_until_end``, ``_with_save``, and ``_save_gather_errors``.
+Mixed into ``BaseScraper`` via MRO — expects ``self.verbose``, ``self.saver``,
+``self.error_count``, ``self.max_workers``, ``self.name``, and
+``self._save_doc_result`` from the host class.
 """
 
 from __future__ import annotations
 
-from typing import TYPE_CHECKING
+import asyncio
+from typing import TYPE_CHECKING, Any, Callable, Coroutine
+
+from loguru import logger
+from tqdm import tqdm
+
+from src.scraper.base.summary_utils import flatten_results, merge_context
 
 if TYPE_CHECKING:
     pass
 
-from typing import Any, Callable, Coroutine
 
+class PaginationMixin:
+    """Pagination and gathering helpers mixed into BaseScraper."""
 
-def _flatten(results: list) -> list:
-    """Flatten a list of lists/items, filtering out None."""
-    flat: list = []
-    for item in results:
-        if isinstance(item, list):
-            flat.extend(item)
-        elif item is not None:
-            flat.append(item)
-    return flat
+    async def _with_save(self, coro, context: dict):
+        result = await coro
+        if result is None:
+            return None
 
+        is_list = isinstance(result, list)
+        items = result if is_list else [result]
+        saved = []
+        for r in items:
+            doc = merge_context(r, context)
+            s = await self._save_doc_result(doc)
+            if s is None:
+                logger.warning(
+                    f"Save failed for '{doc.get('title', '?')}', discarding result"
+                )
+                continue
+            saved.append(s)
+        if not saved:
+            return None
+        return saved if is_list else saved[0]
 
-class PaginationHelper:
-    """Pagination utilities composed into BaseScraper as ``self.paginator``.
+    async def _save_gather_errors(
+        self,
+        results: list,
+        context: dict,
+        desc: str = "",
+    ) -> list:
+        ctx = {"year": "", "type": "", "situation": "", **context}
+        valid = []
+        for result in results:
+            if isinstance(result, Exception):
+                self.error_count += 1
+                logger.error(f"{desc} | Error: {result}")
+                if self.saver:
+                    error_data = {
+                        "title": desc or "Unknown",
+                        "html_link": "",
+                        **ctx,
+                    }
+                    await self.saver.save_error(error_data, error_message=str(result))
+                continue
+            if result is None:
+                continue
+            valid.append(result)
+        return valid
 
-    ``gather_fn`` must be ``BaseScraper._gather_results`` (bound method).
-    """
+    async def _gather_results(
+        self,
+        tasks: list,
+        context: dict | None = None,
+        desc: str = "",
+    ) -> list:
+        if not tasks:
+            return []
 
-    def __init__(self, gather_fn, max_workers: int):
-        self._gather = gather_fn
-        self.max_workers = max_workers
+        if self.verbose:
+            progress = tqdm(total=len(tasks), desc=desc or "Gathering")
+            wrapped_tasks = [asyncio.create_task(task) for task in tasks]
+            for task in wrapped_tasks:
+                task.add_done_callback(lambda _: progress.update())
+            try:
+                results = await asyncio.gather(*wrapped_tasks, return_exceptions=True)
+            finally:
+                progress.close()
+        else:
+            results = await asyncio.gather(*tasks, return_exceptions=True)
+        return await self._save_gather_errors(results, context or {}, desc)
 
-    def calc_pages(self, total: int, per_page: int) -> int:
-        """Number of pages needed for *total* items at *per_page* each."""
-        if total <= 0 or per_page <= 0:
-            return 0
-        return (total + per_page - 1) // per_page
+    async def _process_documents(
+        self,
+        documents: list,
+        *,
+        year: int,
+        norm_type: str,
+        situation: str = "",
+        desc: str = "",
+        doc_data_fn=None,
+        doc_data_kwargs: dict | None = None,
+    ) -> list[dict]:
+        """Wrap each document through _get_doc_data -> _with_save -> _gather_results."""
+        ctx = {"year": year, "type": norm_type, "situation": situation}
+        fn = doc_data_fn or self._get_doc_data
+        kw = doc_data_kwargs or {}
+        # Enrich each doc with context fields (year/type/situation) so that
+        # ScrapedDocument validation succeeds when _get_doc_data calls
+        # _process_pdf_doc/_process_html_doc.  Doc-specific values take
+        # precedence over context defaults (dict merge order: ctx first, then doc).
+        enriched = [{**ctx, **doc} for doc in documents]
+        tasks = [self._with_save(fn(doc, **kw), ctx) for doc in enriched]
+        results = await self._gather_results(
+            tasks,
+            context=ctx,
+            desc=desc or f"{self.name} | {norm_type}",
+        )
+        logger.debug(
+            f"Finished scraping for Year: {year} | Type: {norm_type} "
+            f"| Situation: {situation} | Results: {len(results)}"
+        )
+        return results
 
-    async def fetch_all_pages(
+    async def _fetch_all_pages(
         self,
         make_task: Callable[[int], Coroutine],
         total_pages: int,
@@ -49,51 +133,14 @@ class PaginationHelper:
         context: dict | None = None,
         desc: str = "",
     ) -> list:
-        """Fetch pages ``start_page``..``total_pages`` concurrently and flatten.
-
-        Typical usage — call after fetching and parsing page 1 yourself::
-
-            docs = first_page_docs
-            extra = await self.paginator.fetch_all_pages(
-                lambda p: self._get_docs_links(self._build_url(year, p)),
-                total_pages,
-                context=ctx,
-                desc="SCRAPER | year | get_docs_links",
-            )
-            docs.extend(extra)
-
-        Returns a flat list of all items gathered from the extra pages.
-        """
+        """Fetch pages ``start_page``..``total_pages`` concurrently and flatten."""
         if total_pages < start_page:
             return []
         tasks = [make_task(page) for page in range(start_page, total_pages + 1)]
-        results = await self._gather(tasks, context=context, desc=desc)
-        return _flatten(results)
+        results = await self._gather_results(tasks, context=context, desc=desc)
+        return flatten_results(results)
 
-    async def collect_paginated_listing(
-        self,
-        first_page_docs: list[dict],
-        *,
-        total_pages: int,
-        make_page_task: Callable[[int], Coroutine],
-        context: dict | None = None,
-        start_page: int = 2,
-        desc: str = "",
-    ) -> list[dict]:
-        """Return first-page docs plus every remaining page in one flat list."""
-        documents = list(first_page_docs)
-        documents.extend(
-            await self.fetch_all_pages(
-                make_page_task,
-                total_pages,
-                start_page=start_page,
-                context=context,
-                desc=desc,
-            )
-        )
-        return documents
-
-    async def paginate_until_end(
+    async def _paginate_until_end(
         self,
         *,
         make_task: Callable[[int], Coroutine[Any, Any, tuple[list[dict], bool]]],
@@ -102,20 +149,19 @@ class PaginationHelper:
         initial_batch: int = 1,
         batch_growth: int | None = None,
         max_batch: int | None = None,
+        max_iterations: int = 1000,
     ) -> list[dict]:
-        """Fetch pages in growing batches until a page signals end-of-results.
-
-        ``make_task(page_number)`` must return ``(docs, reached_end)``.
-        """
+        """Fetch pages in growing batches until a page signals end-of-results."""
         batch = initial_batch
         growth = batch_growth if batch_growth is not None else self.max_workers
         cap = max_batch or self.max_workers
         page = 1
         all_docs: list[dict] = []
+        iterations = 0
 
-        while True:
+        while iterations < max_iterations:
             tasks = [make_task(p) for p in range(page, page + batch)]
-            results = await self._gather(tasks, context=context, desc=desc)
+            results = await self._gather_results(tasks, context=context, desc=desc)
 
             reached_end = False
             batch_docs: list[dict] = []
@@ -131,5 +177,11 @@ class PaginationHelper:
 
             page += batch
             batch = min(batch + growth, cap)
+            iterations += 1
+        else:
+            logger.warning(
+                f"_paginate_until_end: reached max_iterations={max_iterations} "
+                f"without end signal — stopping ({desc})"
+            )
 
         return all_docs
